JSONParserTest.java

import static org.junit.Assert.*;
import org.junit.Before;
import org.junit.Test;
import java.util.List;

public class JSONParserTest {

    private JSONParser parser;

    @Before
    public void setUp() {
        parser = new JSONParser();
    }

    @Test
    public void testParseValidJSON() {
        // Assuming a sample JSON file "test_data.json" with valid credit card data
        List<CreditCard> cards = parser.parseInputFile("input_file.json");
        assertNotNull(cards);
        assertFalse(cards.isEmpty());

        // Further assertions can be added as required
    }

    // Additional tests for different scenarios can be implemented here
}
